Add RegisterNode endpoint and functionality
Add RegisterNode endpoint to `Handler` interface in `handler.go`.
Implement the RegisterNode function in `handler.go`.

Add RegisterNodeRequest struct definition to `requests.go`.

Add NewHost and NewNode methods to `nodeapi.go`.
Implement NewNode method to add the new host to `knownHosts` and call makePeers to update peers.

Implement the NewNode method in `service.go` to add the new node's address to the known hosts and call the `NewHost` method.

The RegisterNode endpoint allows registering a new node in the gateway.

Signed-off-by: Yuriy Kis <yurakis9@gmail.com>