feat: Add API listen address and gateway address configuration
This commit adds the configuration for the API listen address and the gateway address in the `docker-compose.yml` file. The API listen address is set to `nodeX:400X`, where `X` is the node number. The gateway address is set to `http://gateway:6000`.

In the `main.go` file, the `GATEWAY_ADDR` environment variable is added and read. If it is empty, the default value is set to `http://localhost:6000`. The `service.New` function is called with the additional `gatewayAddress` parameter.

In the `gateway.go` file, the error and success log messages now include the endpoint address of the gateway.

In the `service.go` file, the `service.New` function is modified to include the `gatewayAddress` parameter for initializing the `gatewayClient`.

Signed-off-by: Yuriy Kis <yurakis9@gmail.com>