Add GRPC metrics instrumentation and refactor transport server implementation
Added a new struct `GRPCMetricsHandler` to handle metrics for GRPC requests. Implemented the `newGRPCMetricsHandler` function to create a new instance of `GRPCMetricsHandler`. Also implemented the `instrument` method in `GRPCMetricsHandler` to instrument the GRPC requests by incrementing the request counter, measuring the latency, and incrementing the error counter if there is an error.

Refactored the `Start` and `Stop` methods of `GRPCNodeServer` to `startGRPCTransport` and `stopGRPCTransport`, respectively. These methods now receive a generic `NodeServer` interface as an argument and cast it to `GRPCNodeServer`. This allows for easier testing and extensibility.

Updated the `node` struct to use the new `NodeServer` interface instead of `TransportServer`. The `transportServer` field has been replaced with `nodeServer`.

Signed-off-by: Yuriy Kis <yurakis9@gmail.com>