feat: Add Consul service for service discovery
This commit adds support for Consul service discovery. It includes changes to the Makefile to add a new target 'clear', changes to the docker-compose.yml file to include consul-server and consul-client services, the addition of client.json and server.json files for consul configurations, and the addition of consul.go file for the implementation of the ConsulService struct in the node package.

Note: The Consul service is started automatically when the node starts.
Signed-off-by: Yuriy Kis <yurakis9@gmail.com>