Makefile: Add consul target to start Consul server and client containers
The Makefile was modified to add a new 'consul' target that starts the Consul server and client containers. This target runs the 'docker compose start' command to start both containers.

consul.go: Remove acceptLoop function from ConsulService

The acceptLoop function in the ConsulService struct was removed from the consul.go file. This function was responsible for accepting connections, but it was no longer being used. The function has been removed to clean up the code.

Signed-off-by: Yuriy Kis <yurakis9@gmail.com>