feat: Add 'up-all' command to start all services at once
This commit introduces a new 'up-all' command in the Makefile to start all services (gateway, node1, node2, node3, node4) using Docker Compose. This makes it easier to bring up all services in a single command.

Additionally, in the 'node/main.go' file, the code has been updated to read the environment variable 'IS_MINER' as a string and convert it to a boolean value using strconv.ParseBool. This change ensures that the 'isMiner' variable is correctly set based on the environment variable value, preventing any misinterpretation.

Signed-off-by: Yuriy Kis <yurakis9@gmail.com>