feat: Add KafkaMessageProducer for producing messages
This commit adds a new file `producer.go` that implements the `MessageProducer` interface to produce messages using the Confluent Kafka Go library. The `KafkaMessageProducer` struct has a `producer` field of type `kafka.Producer` that is used to produce messages.

The `ProduceMessage` method of the `KafkaMessageProducer` takes a message object as input, marshals it into JSON, determines the topic based on the type of message, and then uses the producer to send the message to Kafka.

In the `gateway.go` file, the `NewGatewayClient` function is modified to create an instance of `KafkaMessageProducer` and assign it to the `mp` field of `gatewayClient`. The `RegisterMe` method is updated to use the `ProduceMessage` method of `mp` to send a `RegisterNodeMessage` to Kafka.

Note: The Kafka broker address is currently hardcoded as `localhost` in the `NewKafkaMessageProducer` function.
Signed-off-by: Yuriy Kis <yurakis9@gmail.com>